fix a few compiler warnings in VC
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@1561 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/water.c
@@ -388,7 +388,7 @@ void SineBlob(int x, int y, int radius, int height, int page)
       square = cy*cy + cx*cx;
       if(square < radsquare)
       {
-        dist = sqrt(square*length);
+        dist = (int) sqrt(square*length);
         Height[page][SCRW*(cy+y) + cx+x]
           += (int)((FCos(dist)+0xffff)*(height)) >> 19;
       }
